+++ api/src/users/users.controller.ts
@@ -13,7 +13,7 @@ import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UserDto } from './dto/user.dto';
-import { AuthGuard } from 'src/auth/guards/auth.guard';
+import { AuthGuard } from '../auth/guards/auth.guard';
 import { AppClient } from '../common/decorators/app-client.decorator';
 
 @Controller('users')
@@ -27,8 +27,8 @@ export class UsersController {
   }
 
   @Get()
-  findAll(): Promise<UserDto[]> {
-    return this.usersService.findAll();
+  findAll(@AppClient() user): Promise<UserDto[]> {
+    return this.usersService.findAll(user.id);
   }
 
   @Get('/me')
@@ -37,7 +37,10 @@ export class UsersController {
   }
 
   @Get(':id')
-  async findOne(@Param('id') id: string): Promise<UserDto> {
+  async findOne(@AppClient() client, @Param('id') id: string): Promise<UserDto> {
+    if (id !== client.id) {
+      throw new HttpException('Acceso denegado', 403);
+    }
     const user = await this.usersService.findOne(id);
     if (!user) {
       throw new HttpException('Usuario no encontrado', 404);
@@ -46,12 +49,18 @@ export class UsersController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  update(@AppClient() client, @Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+    if (id !== client.id) {
+      throw new HttpException('Acceso denegado', 403);
+    }
     return this.usersService.update(id, updateUserDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@AppClient() client, @Param('id') id: string) {
+    if (id !== client.id) {
+      throw new HttpException('Acceso denegado', 403);
+    }
     return this.usersService.remove(id);
   }
 }
